Skin Disease/requirements.txt
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
opencv-python==4.8.1.78
scipy==1.11.2
scikit-image==0.21.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
werkzeug==2.3.7
efficientnet==1.0.0
biopython==1.81
huggingface-hub==0.17.2
datasets==2.14.4
